llm/requirements.txt
langchain
wikipedia
langchain-community
python-dotenv
pydantic
duckduckgo-search
langchain_huggingface
streamlit
langchain-nvidia-ai-endpoints
langchain-ollama
langchain-chroma
# Document processing
PyPDF2
python-docx
email-validator
# Vector database and embeddings
faiss-cpu
sentence-transformers
# Additional utilities
tiktoken
numpy
pandas
# For structured outputs
pydantic
# For API endpoints
fastapi
uvicorn
python-multipart